paho_example/requirements.txt
paho-mqtt>=1.4.0
urllib3>=1.24.1
